+++ zarrs/src/array/codec/array_to_bytes/codec_chain.rs
@@ -9,9 +9,9 @@ use crate::{
         array_bytes::update_bytes_flen,
         codec::{
             ArrayCodecTraits, ArrayPartialDecoderCache, ArrayPartialDecoderTraits,
-            ArrayToArrayCodecTraits, ArrayToBytesCodecTraits, BytesPartialDecoderCache,
-            BytesPartialDecoderTraits, BytesToBytesCodecTraits, Codec, CodecError, CodecOptions,
-            CodecTraits,
+            ArrayPartialEncoderTraits, ArrayToArrayCodecTraits, ArrayToBytesCodecTraits,
+            BytesPartialDecoderCache, BytesPartialDecoderTraits, BytesPartialEncoderTraits,
+            BytesToBytesCodecTraits, Codec, CodecError, CodecOptions, CodecTraits,
         },
         concurrency::RecommendedConcurrency,
         ArrayBytes, ArrayMetadataOptions, BytesRepresentation, ChunkRepresentation, ChunkShape,
@@ -451,6 +451,74 @@ impl ArrayToBytesCodecTraits for CodecChain {
         }
 
         Ok(input_handle)
+    }
+
+    fn partial_encoder(
+        self: Arc<Self>,
+        mut input_handle: Arc<dyn BytesPartialDecoderTraits>,
+        mut output_handle: Arc<dyn BytesPartialEncoderTraits>,
+        decoded_representation: &ChunkRepresentation,
+        options: &CodecOptions,
+    ) -> Result<Arc<dyn ArrayPartialEncoderTraits>, CodecError> {
+        let array_representations =
+            self.get_array_representations(decoded_representation.clone())?;
+        let bytes_representations =
+            self.get_bytes_representations(array_representations.last().unwrap())?;
+
+        for (codec, bytes_representation) in std::iter::zip(
+            self.bytes_to_bytes.iter().rev(),
+            bytes_representations.iter().rev().skip(1),
+        ) {
+            output_handle = Arc::clone(codec).partial_encoder(
+                input_handle.clone(),
+                output_handle,
+                bytes_representation,
+                options,
+            )?;
+            input_handle =
+                Arc::clone(codec).partial_decoder(input_handle, bytes_representation, options)?;
+        }
+
+        let mut output_handle = self.array_to_bytes.clone().partial_encoder(
+            input_handle.clone(),
+            output_handle,
+            array_representations.last().unwrap(),
+            options,
+        )?;
+
+        if self.array_to_array.is_empty() {
+            return Ok(output_handle);
+        }
+
+        let mut input_handle = self.array_to_bytes.clone().partial_decoder(
+            input_handle,
+            array_representations.last().unwrap(),
+            options,
+        )?;
+
+        let mut it = std::iter::zip(
+            self.array_to_array.iter().rev(),
+            array_representations.iter().rev().skip(1),
+        )
+        .peekable();
+        while let Some((codec, array_representation)) = it.next() {
+            output_handle = Arc::clone(codec).partial_encoder(
+                input_handle.clone(),
+                output_handle,
+                array_representation,
+                options,
+            )?;
+
+            if it.peek().is_some() {
+                input_handle = Arc::clone(codec).partial_decoder(
+                    input_handle,
+                    array_representation,
+                    options,
+                )?;
+            }
+        }
+
+        Ok(output_handle)
     }
 
     #[cfg(feature = "async")]